+++ config/msegat.php
@@ -2,7 +2,7 @@
 
 // config for BitcodeSa/Msegat
 return [
-    "api_url" => env("MSEGAT_API_URL", "https://www.msegat.com/gw/sendsms.php"),
+    "api_url" => env("MSEGAT_API_URL", "https://www.msegat.com/gw"),
     "api_key" => env("MSEGAT_API_KEY", ""),
     "username" => env("MSEGAT_USERNAME", ""),
     "sender" => env("MSEGAT_SENDER", ""),