+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # Changelog
+
+### Version 1.0.1:
+- Add ability to request otp.
+- update to the Msegat class structure.